refactor(performance): keep the federation declarations off the public API
CI reported `types.d.ts need to be updated`: exposing the accessors as
statics on two exported plugins pulled a generated `SharedDeclaration`
interface into the public declarations, and hand-splicing that back is
version-sensitive — the repo's generator and CI's disagree.

The declarations are plumbing between a producer and one reader, so they
belong in a module neither exports. `lib/sharing/declaredShared.js` and
`lib/container/declaredRemotes.js` now hold the WeakMap and the add/get
pair, which leaves `types.d.ts` differing from main by only the schema
description this pull request already changed.
