+++ lib/container/ContainerReferencePlugin.js
@@ -14,6 +14,7 @@ const FallbackModuleFactory = require("./FallbackModuleFactory");
 const RemoteModule = require("./RemoteModule");
 const RemoteRuntimeModule = require("./RemoteRuntimeModule");
 const RemoteToExternalDependency = require("./RemoteToExternalDependency");
+const { addDeclaredRemotes } = require("./declaredRemotes");
 const { parseOptions } = require("./options");
 
 /**
@@ -21,22 +22,10 @@ const { parseOptions } = require("./options");
  * 	ContainerReferencePluginOptions
  * } from "../../declarations/plugins/container/ContainerReferencePlugin"
  */
-/** @import Compilation from "../Compilation" */
 /** @import Compiler from "../Compiler" */
 
 const slashCode = "/".charCodeAt(0);
 const PLUGIN_NAME = "ContainerReferencePlugin";
-
-/** @type {WeakMap<Compilation, Set<string>>} */
-const DECLARED_REMOTES = new WeakMap();
-
-/**
- * The remote names this build declared, for `performance.unusedConfig` to tell
- * apart from the ones a module actually imported.
- * @param {Compilation} compilation the compilation
- * @returns {Set<string> | undefined} the declared names, when any were declared
- */
-const getDeclaredRemotes = (compilation) => DECLARED_REMOTES.get(compilation);
 
 class ContainerReferencePlugin {
 	/**
@@ -103,14 +92,10 @@ class ContainerReferencePlugin {
 		compiler.hooks.compilation.tap(
 			PLUGIN_NAME,
 			(compilation, { normalModuleFactory }) => {
-				let declared = DECLARED_REMOTES.get(compilation);
-
-				if (declared === undefined) {
-					declared = new Set();
-					DECLARED_REMOTES.set(compilation, declared);
-				}
-
-				for (const [key] of remotes) declared.add(key);
+				addDeclaredRemotes(
+					compilation,
+					remotes.map(([key]) => key)
+				);
 
 				compilation.dependencyFactories.set(
 					RemoteToExternalDependency,
@@ -177,7 +162,5 @@ class ContainerReferencePlugin {
 		);
 	}
 }
-
-ContainerReferencePlugin.getDeclaredRemotes = getDeclaredRemotes;
 
 module.exports = ContainerReferencePlugin;
