+++ lib/performance/UnusedFederationPlugin.js
@@ -5,11 +5,11 @@
 
 "use strict";
 
-const { getDeclaredRemotes } = require("../container/ContainerReferencePlugin");
 const RemoteModule = require("../container/RemoteModule");
+const { getDeclaredRemotes } = require("../container/declaredRemotes");
 const UnusedFederationWarning = require("../errors/UnusedFederationWarning");
 const ConsumeSharedModule = require("../sharing/ConsumeSharedModule");
-const { getDeclaredShared } = require("../sharing/ConsumeSharedPlugin");
+const { getDeclaredShared } = require("../sharing/declaredShared");
 const { compareStrings } = require("../util/comparators");
 
 /** @import { PerformanceOptions } from "../../declarations/WebpackOptions" */
