+++ lib/sharing/ConsumeSharedPlugin.js
@@ -15,6 +15,7 @@ const ConsumeSharedFallbackDependency = require("./ConsumeSharedFallbackDependen
 const ConsumeSharedModule = require("./ConsumeSharedModule");
 const ConsumeSharedRuntimeModule = require("./ConsumeSharedRuntimeModule");
 const ProvideForSharedDependency = require("./ProvideForSharedDependency");
+const { addDeclaredShared } = require("./declaredShared");
 const { resolveMatchedConfigs } = require("./resolveMatchedConfigs");
 const {
 	getDescriptionFile,
@@ -32,7 +33,6 @@ const getModuleNotFoundError = memoize(() =>
  * 	ConsumeSharedPluginOptions
  * } from "../../declarations/plugins/sharing/ConsumeSharedPlugin"
  */
-/** @import Compilation from "../Compilation" */
 /** @import Compiler from "../Compiler" */
 /** @import { FileSystemDependencies } from "../Compilation" */
 /** @import { ResolveOptionsWithDependencyType } from "../ResolverFactory" */
@@ -43,25 +43,6 @@ const getModuleNotFoundError = memoize(() =>
 /** @type {ResolveOptionsWithDependencyType} */
 const RESOLVE_OPTIONS = { dependencyType: "esm" };
 const PLUGIN_NAME = "ConsumeSharedPlugin";
-
-/**
- * @typedef {object} SharedDeclaration
- * @property {string} name the config key, as the config spells it
- * @property {string} shareKey the key the modules created for it carry
- * @property {boolean} prefix whether it shares everything under the key
- */
-
-/** @type {WeakMap<Compilation, SharedDeclaration[]>} */
-const DECLARED_SHARED = new WeakMap();
-
-/**
- * What this build declared as shared, one entry per config key, since several
- * may name one share key. `performance.unusedConfig` matches modules on the
- * share key and reports the config key.
- * @param {Compilation} compilation the compilation
- * @returns {SharedDeclaration[] | undefined} the declarations, when any were made
- */
-const getDeclaredShared = (compilation) => DECLARED_SHARED.get(compilation);
 
 class ConsumeSharedPlugin {
 	/**
@@ -157,22 +138,16 @@ class ConsumeSharedPlugin {
 					normalModuleFactory
 				);
 
-				let declared = DECLARED_SHARED.get(compilation);
-
-				if (declared === undefined) {
-					declared = [];
-					DECLARED_SHARED.set(compilation, declared);
-				}
-
-				for (const [key, config] of consumes) {
-					// A key ending in a slash shares everything under it, and the
-					// modules then carry the remainder appended to the share key.
-					declared.push({
+				// A key ending in a slash shares everything under it, and the
+				// modules then carry the remainder appended to the share key.
+				addDeclaredShared(
+					compilation,
+					consumes.map(([key, config]) => ({
 						name: key,
 						shareKey: config.shareKey,
 						prefix: key.endsWith("/")
-					});
-				}
+					}))
+				);
 
 				/** @typedef {Map<string, ConsumeOptions>} Consumes */
 
@@ -428,7 +403,5 @@ class ConsumeSharedPlugin {
 		);
 	}
 }
-
-ConsumeSharedPlugin.getDeclaredShared = getDeclaredShared;
 
 module.exports = ConsumeSharedPlugin;
